+++ samples/application/gulpfile.js
@@ -1,0 +1,14 @@
+// Add standard tasks
+var gulp = require('../../src/index').gulp;
+require('../../src/index').all(gulp);
+
+// Define build tasks        
+gulp.task('build', ['build-dev', 'build-prod']);
+gulp.task('rebuild', ['build-dev']);
+gulp.task('clean', ['build-clean']);
+gulp.task('watch', ['build-watch']);
+gulp.task('jshint', ['test-jshint']);
+gulp.task('launch', ['app-launch']);
+
+// Set default task
+gulp.task('default', ['build']);