+++ samples/module/src/core.js
@@ -1,0 +1,17 @@
+﻿/**
+ * @file Registration of all WebUI core services
+ * @copyright Digital Living Software Corp. 2014-2016
+ */
+
+/* global angular */
+
+(function () {
+    'use strict';
+
+    angular.module('pipCore', [
+	    'pipTranslate',
+        'pipAssert',
+        'pipDebug'
+    ]);
+    
+})();