+++ src/samples.js
@@ -1,9 +1,8 @@
 var rename = require('gulp-rename');
 var replace = require('gulp-replace');
 var es = require('event-stream');
-var awspublish = require('gulp-awspublish');
 var webserver = require('gulp-webserver');
-var parallelize = require('concurrent-transform');
+var publish = require('./publish');
 
 var pkg = require(process.cwd() + '/package.json');
 var conf = require('./config');
@@ -12,14 +11,7 @@ module.exports = function (gulp) {
 
     gulp.task('samples-publish', function () {
         var
-            publisher = awspublish.create({
-                params: {
-                    Bucket: conf.samples.publish.bucket
-                },
-                accessKeyId: conf.samples.publish.accessKeyId,
-                secretAccessKey: conf.samples.publish.secretAccessKey,
-                region: conf.samples.publish.region
-            }),
+            storage = conf.samples.publish,
 
             distFiles = gulp.src([
                 conf.dir.lib + '/**/*',
@@ -33,25 +25,28 @@ module.exports = function (gulp) {
                 .pipe(replace(pkg.name + '-lib.js', pkg.name + '-lib.min.js'))
                 .pipe(replace(pkg.name + '.css', pkg.name + '.min.css'))
                 .pipe(replace(pkg.name + '.js', pkg.name + '.min.js'))
-                .pipe(replace('../../.' + conf.dir.lib, '/' + pkg.name + '/'))
-                .pipe(replace('../.' + conf.dir.lib, '/' + pkg.name + '/'))
-                .pipe(replace('.' + conf.dir.lib, '/' + pkg.name + '/'))
-                .pipe(replace('../../.' + conf.dir.dist, '/' + pkg.name + '/'))
-                .pipe(replace('../.' + conf.dir.dist, '/' + pkg.name + '/'))
-                .pipe(replace('.' + conf.dir.dist, '/' + pkg.name + '/')),
+                .pipe(replace('../../.' + conf.dir.lib, '')) // '/' + pkg.name + '/'))
+                .pipe(replace('../.' + conf.dir.lib, '')) // '/' + pkg.name + '/'))
+                .pipe(replace('.' + conf.dir.lib, '')) // '/' + pkg.name + '/'))
+                .pipe(replace('../../.' + conf.dir.dist, '')) // '/' + pkg.name + '/'))
+                .pipe(replace('../.' + conf.dir.dist, '')) // '/' + pkg.name + '/'))
+                .pipe(replace('.' + conf.dir.dist, '')), // '/' + pkg.name + '/')),
 
             sampleFiles = gulp.src([
                 conf.dir.samples + '**/*',
                 '!' + conf.dir.samples + '**/index.*'
             ], {xbase: '.'});
 
-        return es.merge([distFiles, indexFiles, sampleFiles])
-            .pipe(rename(function (path) {
-                path.dirname = '/' + pkg.name + '/' + path.dirname
-            }))
-            .pipe(parallelize(publisher.publish(), 5))
-            .pipe(publisher.sync(pkg.name))
-            .pipe(awspublish.reporter());
+        var 
+            files = es.merge([distFiles, indexFiles, sampleFiles])
+                .pipe(rename(function (path) {
+                    path.dirname = '/' + pkg.name + '/' + path.dirname
+                }));
+
+        // Put sample into folder same as module name
+        storage.folder = storage.folder || pkg.name;
+
+        return publish(storage, files);
     });
 
     gulp.task('samples-launch', function () {